Move CheckVolumeDir function
