+++ internal/pkg/service/buffer/storage/local/volume/check.go
@@ -1,12 +1,10 @@
-package local
+package volume
 
 import (
 	"os"
 
 	"github.com/keboola/keboola-as-code/internal/pkg/utils/errors"
 )
-
-const VolumeIDFile = "volume-id.txt"
 
 func CheckVolumeDir(path string) error {
 	info, err := os.Stat(path)
