+++ internal/pkg/service/buffer/storage/local/volume/check_test.go
@@ -1,4 +1,4 @@
-package local
+package volume
 
 import (
 	"fmt"
